проработаны работа с категориями

+++ main-service/main-service-admin/src/main/java/ru/yandex/practicum/ewm/categories/service/CategoryServiceImpl.java
@@ -1,20 +1,16 @@
 package ru.yandex.practicum.ewm.categories.service;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.data.crossstore.ChangeSetPersister;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+
 import ru.yandex.practicum.ewm.categories.dto.CategoryDto;
 import ru.yandex.practicum.ewm.categories.dto.NewCategoryDto;
 import ru.yandex.practicum.ewm.categories.mapper.CategoryMapper;
+import ru.yandex.practicum.ewm.exception.exceptionType.NotFoundException;
+import ru.yandex.practicum.ewm.exception.exceptionType.ValidationException;
 import ru.yandex.practicum.ewm.model.Category;
 import ru.yandex.practicum.ewm.repository.CategoryRepository;
 
-import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,34 +19,35 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDto addCategory(NewCategoryDto dto) {
-        /*  Category category = CategoryMapper.toEntity(dto);
-        try {
-            Category saved = categoryRepository.save(category);
-            return CategoryMapper.toDto(saved);
-        } catch (DataIntegrityViolationException e) {
-            throw new ConflictException("Category name already exists: " + dto.getName());
-        }*/
-        return  null;
+        if (categoryRepository.existsByName(dto.getName())) {
+            throw new ValidationException("Категория с таким именем уже существует");
+        }
+        Category category = CategoryMapper.toEntity(dto);
+        Category saved = categoryRepository.save(category);
+        return CategoryMapper.toDto(saved);
     }
 
     @Override
     public CategoryDto updateCategory(Long catId, CategoryDto dto) {
-      /*  Optional<Category> category = categoryRepository.findById(catId);
-      //  category.setName(dto.getName());
-
-
-       // Category updated = categoryRepository.save(category);
-        return CategoryMapper.toDto(updated);*/
-        return null;
+        Category category = categoryRepository.findById(catId)
+                .orElseThrow(() ->
+                        new NotFoundException("Категория не найдена"));
+        if (categoryRepository.existsByName(dto.getName())) {
+            throw new ValidationException("Категория с таким именем уже существует");
+        }
+        category.setName(dto.getName());
+        Category updated = categoryRepository.save(category);
+        return CategoryMapper.toDto(updated);
 
     }
 
     @Override
     public void deleteCategory(Long catId) {
+        if (!categoryRepository.existsById(catId)) {
+            throw new NotFoundException("Категория не найдена");
+        }
         categoryRepository.deleteById(catId);
-
     }
-
 
 
 }
